requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   smsgate (pyproject.toml)
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
async-lru==2.0.5
    # via smsgate (pyproject.toml)
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
coverage==7.9.1
    # via pytest-cov
diskcache==5.6.3
    # via smsgate (pyproject.toml)
fastapi==0.115.12
    # via smsgate (pyproject.toml)
google-auth==2.40.3
    # via google-genai
google-genai==1.20.0
    # via smsgate (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   google-genai
    #   pocketbase
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
msgpack==1.1.1
    # via smsgate (pyproject.toml)
nats-py==2.10.0
    # via smsgate (pyproject.toml)
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pocketbase==0.15.0
    # via smsgate (pyproject.toml)
prometheus-client==0.22.1
    # via smsgate (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.6
    # via
    #   fastapi
    #   google-genai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via smsgate (pyproject.toml)
pygments==2.19.1
    # via pytest
pytest==8.4.0
    # via
    #   smsgate (pyproject.toml)
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
    #   pytest-tornasync
pytest-asyncio==1.0.0
    # via smsgate (pyproject.toml)
pytest-cov==6.2.1
    # via smsgate (pyproject.toml)
pytest-mock==3.14.1
    # via smsgate (pyproject.toml)
pytest-tornasync==0.6.0.post2
    # via smsgate (pyproject.toml)
python-dateutil==2.9.0.post0
    # via smsgate (pyproject.toml)
python-dotenv==1.1.0
    # via
    #   smsgate (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
redis==6.2.0
    # via smsgate (pyproject.toml)
requests==2.32.4
    # via
    #   smsgate (pyproject.toml)
    #   google-genai
rsa==4.9.1
    # via google-auth
sentry-sdk==2.30.0
    # via smsgate (pyproject.toml)
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via smsgate (pyproject.toml)
tornado==6.5.1
    # via pytest-tornasync
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.34.3
    # via smsgate (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchdog==6.0.0
    # via smsgate (pyproject.toml)
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
